pyproject.toml
[tool.poetry]
name = "ossindex-lib"
version = "0.2.1"
description = "A library for querying the OSS Index free catalogue of open source components to help developers identify vulnerabilities, understand risk, and keep their software safe."
authors = ["Paul Horton <phorton@sonatype.com>"]
maintainers = ["Paul Horton <phorton@sonatype.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/sonatype-nexus-community/ossindex-python"
repository = "https://github.com/sonatype-nexus-community/ossindex-python"
packages = [
    { include = "ossindex" }
]
include = [
    "LICENSE"
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'Intended Audience :: Information Technology',
    'Intended Audience :: Legal Industry',
    'Intended Audience :: System Administrators',
    'Topic :: Security',
    'Topic :: Software Development',
    'Topic :: System :: Software Distribution',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Typing :: Typed'
]
keywords = [
    "SCA", "OWASP"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/sonatype-nexus-community/ossindex-python/issues"

[tool.poetry.dependencies]
python = "^3.6"
importlib-metadata = { version = ">= 3.4", python = "< 3.8" }
packageurl-python = "^0.9.0"
PyYAML = "^5.4.1"
requests = "^2.20.0"
tinydb = "^4.5.0"
# `types-requeests` should stay in sync with `requests`
types-requests = "^2.25.1"
# `types-setuptools` need to stay in sync with version of `setuptools` - but 47 was not typed...
types-setuptools = ">= 57.0.0"

[tool.poetry.dev-dependencies]
tox = "^3.24.3"
coverage = "^6.2"
flake8 = "^4.0.1"
mypy = ">= 0.920, < 1.00"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "pyproject.toml:version"
]
branch = "master"
upload_to_pypi = true
upload_to_release = true
build_command = "pip install poetry && poetry build"